@@ -1,3 +1,4 @@
+import 'package:flt_imo/Controller/TabViewController/dashboardController.dart';
 import 'package:flt_imo/Models/inventoryModel.dart';
 import 'package:flt_imo/Models/locationModel.dart';
 import 'package:flt_imo/NoInternetConnection/no_internet.dart';
@@ -13,6 +14,7 @@ class AddBoxesController extends GetxController {
 
   var isSaleable = false.obs;
   var isFragile = false.obs;
+  var isDetailLoading = false.obs;
   var isNegotiable = false.obs;
   var processLoading = false.obs;
 
@@ -42,22 +44,22 @@ class AddBoxesController extends GetxController {
       mySnackbar(title: txtRequired, description: txtEnterDescription);
       return false;
     }
-    if (weightController.text == "") {
-      mySnackbar(title: txtRequired, description: txtEnterWeight);
-      return false;
-    }
-    if (heightController.text == "") {
-      mySnackbar(title: txtRequired, description: txtEnterHeight);
-      return false;
-    }
-    if (widthController.text == "") {
-      mySnackbar(title: txtRequired, description: txtEnterWidth);
-      return false;
-    }
-    if (lenghtController.text == "") {
-      mySnackbar(title: txtRequired, description: txtEnterLength);
-      return false;
-    }
+    // if (weightController.text == "") {
+    //   mySnackbar(title: txtRequired, description: txtEnterWeight);
+    //   return false;
+    // }
+    // if (heightController.text == "") {
+    //   mySnackbar(title: txtRequired, description: txtEnterHeight);
+    //   return false;
+    // }
+    // if (widthController.text == "") {
+    //   mySnackbar(title: txtRequired, description: txtEnterWidth);
+    //   return false;
+    // }
+    // if (lenghtController.text == "") {
+    //   mySnackbar(title: txtRequired, description: txtEnterLength);
+    //   return false;
+    // }
     if (priceController.text == "") {
       mySnackbar(title: txtRequired, description: txtEnterPrice);
       return false;
@@ -69,7 +71,7 @@ class AddBoxesController extends GetxController {
     if (await isConnected()) {
       if (validate()) {
         if (flag == 0) {
-          addInventory(getable);
+          addBox(getable);
         } else {
           Get.back();
         }
@@ -93,23 +95,31 @@ class AddBoxesController extends GetxController {
     });
   }
 
-  addInventory(getAble) async {
+  clearParam() {
+    selectedInventory.value = Inventory(name: "Select Inventory");
+    selectedLocation.value = Location(name: "Select Location");
+    titleController.text = "";
+    descriptionController.text = "";
+    isSaleable.value = false;
+    priceController.text = "";
+    isNegotiable.value = false;
+    weightController.text = "";
+    lenghtController.text = "";
+    widthController.text = "";
+    heightController.text = "";
+  }
+
+  addBox(getAble) async {
     processLoading(true);
     var body = setBody();
 
-    addBoxesApi(body).then((response) {
-      if (response == null) {
-        processLoading(false);
-        mySnackbar(title: txtFailed, description: txtUnkownError);
-      }
-      if (response.statusCode == 201) {
-        snackBarBack(title: txtSuccess, description: "Inventory Added Successfully").then((value) {
+    BoxService.addBoxesApi(body).then((response) {
+      if (response != null) {
+        snackBarBack(title: txtSuccess, description: "Box Added Successfully").then((value) {
           Get.back();
+          Get.find<DashboardController>().getRefreshData();
           processLoading(false);
         });
-      } else {
-        processLoading(false);
-        mySnackbar(title: txtFailed, description: txtUnkownError);
       }
     }).catchError((e) {
       printError();
