@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:flt_imo/Bindings/ProfileBinding.dart';
 import 'package:flt_imo/Controller/GeneralController/accessController.dart';
 import 'package:flt_imo/Models/loginResponseModel.dart';
 import 'package:flt_imo/Screen/Auth/login.dart';
@@ -13,18 +14,18 @@ import 'package:get/get.dart';
 
 silentLogin() {
   var body = {"userId": getPrefValue(Keys.USER_ID), "refreshToken": getPrefValue(Keys.REFRESH_TOKEN)};
-  silentLoginApi(body).then((response) {
-    if (response.statusCode == 200) {
+  UserService.silentLoginApi(body).then((response) {
+    if (response != null) {
       var l = loginResponseFromJson(response.body);
-      setPrefValue(Keys.AUTH_TOKEN, l.tokens.idToken.toString());
-      setPrefValue(Keys.ACCESS_TOKEN, l.tokens.accessToken.toString());
+      setPrefValue(Keys.AUTH_TOKEN, l.tokens!.idToken.toString());
+      setPrefValue(Keys.ACCESS_TOKEN, l.tokens!.accessToken.toString());
       // setPrefValue(Keys.REFRESH_TOKEN, l.tokens.refreshToken.toString());
       setPrefValue(Keys.USER_ID, l.userId.toString());
     } else {
       setPrefValue(Keys.AUTH_TOKEN, "");
       setPrefValue(Keys.ACCESS_TOKEN, "");
       setPrefValue(Keys.REFRESH_TOKEN, "");
-      // Get.offAll(Login());
+      Get.offAll(Login());
     }
   });
 }
@@ -48,9 +49,9 @@ class _SplashScreenState extends State<SplashScreen> {
   Widget build(BuildContext context) {
     Future.delayed(Duration(seconds: 2), () {
       if (getPrefValue(Keys.AUTH_TOKEN).isNotEmpty) {
-        Get.offAll(ProjectListScreen());
+        Get.offAll(() => ProjectListScreen(), binding: ProfileBinding());
       } else {
-        Get.offAll(Login());
+        Get.offAll(() => Login());
       }
     });
     return Scaffold(
