@@ -1,136 +1,49 @@
-import 'dart:convert';
+import 'package:flt_imo/APIS/HttpService.dart';
+import 'package:flt_imo/APIS/ServiceConstants.dart';
+import 'package:flt_imo/APIS/httpServe.dart';
 import 'package:flt_imo/Utils/app_constants.dart';
-import 'package:http/http.dart' as http;
-import 'package:flt_imo/Service/constantService.dart';
+import 'package:flt_imo/Utils/mySnackbar.dart';
+import 'package:flt_imo/Utils/strings.dart';
 import 'package:flt_imo/Utils/urls.dart';
+import 'package:get/get.dart';
 
-/*---------------------------Boxes Api's-------------------------- */
-Future<http.Response> boxesListApi() async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      var response = await client.get(
-        BASE_URL + PROJECT + AppConstants.PROJECT.id.toString() + PROJECT_BOXES,
-        headers: {'Authorization': getToken()},
-      );
-      print("Boxes List Response: ${response.request.url}");
-      print("Boxes List Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
+class BoxService {
+  static HttpService httpService = HttpServe();
+
+  static Future<Response?> boxesListApi() async {
+    Response res =
+        await httpService.getRequest(PROJECT + AppConstants.PROJECT.id.toString() + PROJECT_BOXES).onError((error, stackTrace) => emptyRes);
+    if (res.statusCode == 200) {
+      return res;
+    } else if (res.statusCode == 700) {
+      mySnackbar(title: txtFailed, description: txtUnkownError);
+      return null;
+    } else {
+      mySnackbar(title: "Failed", description: "${res.body["message"]}");
     }
-  } else {
-    return emptyRes;
+  }
+
+  static Future<Response?> addBoxesApi(body) async {
+    Response res = await httpService.postRequest(ADD_BOXES, body).onError((error, stackTrace) => emptyRes);
+    if (res.statusCode == 201) {
+      return res;
+    } else if (res.statusCode == 700) {
+      mySnackbar(title: txtFailed, description: txtUnkownError);
+      return null;
+    } else {
+      mySnackbar(title: "Failed", description: "${res.body["message"]}");
+    }
+  }
+
+  static Future<Response?> updateBoxesApi(body, id) async {
+    Response res = await httpService.putRequest(BOXES + id, body).onError((error, stackTrace) => emptyRes);
+    if (res.statusCode == 200) {
+      return res;
+    } else if (res.statusCode == 700) {
+      mySnackbar(title: txtFailed, description: txtUnkownError);
+      return null;
+    } else {
+      mySnackbar(title: "Failed", description: "${res.body["message"]}");
+    }
   }
 }
-
-Future<http.Response> addBoxesApi(body) async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      print("Box Req $body");
-      var response = await client.post(
-        BASE_URL + ADD_BOXES,
-        headers: {'Authorization': getToken(), "accept": "text/plain", "Content-Type": "application/json"},
-        body: jsonEncode(body),
-      );
-      print("Add Boxes Response: ${response.request.url}");
-      print("Add Boxes Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
-    }
-  } else {
-    return emptyRes;
-  }
-}
-
-Future<http.Response> updateBoxesApi(id, body) async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      var response = await client.put(
-        BASE_URL + BOXES + id,
-        headers: {'Authorization': getToken(), "accept": "text/plain", "Content-Type": "application/json"},
-        body: jsonEncode(body),
-      );
-      print("Inventory List Response: ${response.request.url}");
-      print("Inventory List Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
-    }
-  } else {
-    return emptyRes;
-  }
-}
-
-// // ignore: missing_return
-// Future<http.Response> addBoxesApi(body) async {
-//   accessController.checkOnline();
-//   if (accessController.isOnline.value) {
-//     var client = http.Client();
-//     var request = http.MultipartRequest('POST', Uri.parse(BASE_URL + ADD_BOXES));
-//     var response;
-//     try {
-//       // if (uploadFile != null) {
-//       //   final file = await http.MultipartFile.fromPath("file", uploadFile.path);
-//       //   request.files.add(file);
-//       // }
-//       // request.fields['InventoryId'] = body["InventoryId"];
-//       request.fields.addAll(body);
-//       request.headers.addAll({'Authorization': getToken(), "Content-Type": "multipart/form-data"});
-//       var res = await request.send();
-//       response = await http.Response.fromStream(res);
-//       print('Add Boxes ${res.request.url}');
-//       print('Add Boxes ${response.body}');
-//     } catch (e) {
-//       print("Box $e");
-//       return response;
-//     } finally {
-//       client.close();
-//     }
-//     return response;
-//   }
-// }
-
-// // ignore: missing_return
-// Future<dio.Response> addBoxesApi(body) async {
-//   accessController.checkOnline();
-//   if (accessController.isOnline.value) {
-//     // dio.FormData formData = new dio.FormData.fromMap(body);
-//     print(body);
-//     try {
-//       const _extra = <String, dynamic>{};
-
-//       final _result = await d.request<Map<String, dynamic>>(ADD_BOXES,
-//           options: dio.RequestOptions(
-//               method: 'POST',
-//               headers: <String, dynamic>{
-//                 'Authorization': getToken(),
-//                 "accept": "text/plain",
-//                 "Content-Type": "multipart/form-data",
-//               },
-//               extra: _extra,
-//               baseUrl: BASE_URL),
-//           data: body);
-//       print('Add Boxes ${_result.request.path}');
-//       print('Add Boxes ${_result.data}');
-//       return _result;
-//     } on dio.DioError catch (e) {
-//       print(e);
-//     }
-//   }
-// }
