@@ -1,168 +1,83 @@
-import 'dart:convert';
-import 'package:http/http.dart' as http;
-import 'package:flt_imo/Service/constantService.dart';
+import 'package:flt_imo/APIS/HttpService.dart';
+import 'package:flt_imo/APIS/ServiceConstants.dart';
+import 'package:flt_imo/APIS/httpServe.dart';
+import 'package:flt_imo/Utils/mySnackbar.dart';
+import 'package:flt_imo/Utils/strings.dart';
 import 'package:flt_imo/Utils/urls.dart';
+import 'package:get/get.dart';
 
-/*---------------------------Project Api's-------------------------- */
-Future<http.Response> projectListApi() async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      var response = await client.get(
-        BASE_URL + PROJECT,
-        headers: {'Authorization': getToken()},
-      );
-      print("Register Response: ${response.request.url}");
-      print("Register Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
+class ProjectService {
+  static HttpService httpService = HttpServe();
+
+  static Future<Response?> projectListApi() async {
+    Response res = await httpService.getRequest(PROJECT).onError((error, stackTrace) => emptyRes);
+    if (res.statusCode == 200) {
+      return res;
+    } else if (res.statusCode == 700) {
+      mySnackbar(title: txtFailed, description: txtUnkownError);
+      return null;
+    } else {
+      mySnackbar(title: "Failed", description: "${res.body["message"]}");
     }
-  } else {
-    return emptyRes;
   }
-}
 
-Future<http.Response> userStatApi() async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      var response = await client.get(
-        BASE_URL + PROJECT,
-        headers: {'Authorization': getToken()},
-      );
-      print("Register Response: ${response.request.url}");
-      print("Register Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
+  static Future<Response?> projectDetailApi(id) async {
+    Response res = await httpService.getRequest(PROJECT + id.toString()).onError((error, stackTrace) => emptyRes);
+    if (res.statusCode == 200) {
+      return res;
+    } else if (res.statusCode == 700) {
+      mySnackbar(title: txtFailed, description: txtUnkownError);
+      return null;
+    } else {
+      mySnackbar(title: "Failed", description: "${res.body["message"]}");
     }
-  } else {
-    return emptyRes;
   }
-}
 
-Future<http.Response> projectDetailApi(id) async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      var response = await client.get(
-        BASE_URL + PROJECT + id.toString(),
-        headers: {'Authorization': getToken()},
-      );
-      print("Project Response: ${response.request.url}");
-      print("Project Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
+  static Future<Response?> projectStatsApi(id) async {
+    Response res = await httpService.getRequest(PROJECT + id.toString() + PROJECT_STATS).onError((error, stackTrace) => emptyRes);
+    if (res.statusCode == 200) {
+      return res;
+    } else if (res.statusCode == 700) {
+      mySnackbar(title: txtFailed, description: txtUnkownError);
+      return null;
+    } else {
+      mySnackbar(title: "Failed", description: "${res.body["message"]}");
     }
-  } else {
-    return emptyRes;
   }
-}
 
-Future<http.Response> projectStatsApi(id) async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      var response = await client.get(
-        BASE_URL + PROJECT + id.toString() + PROJECT_STATS,
-        headers: {'Authorization': getToken()},
-      );
-      print("Project Response: ${response.request.url}");
-      print("Project Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
+  static Future<Response?> addProjectApi(body) async {
+    Response res = await httpService.postRequest(PROJECT, body).onError((error, stackTrace) => emptyRes);
+    if (res.statusCode == 201) {
+      return res;
+    } else if (res.statusCode == 700) {
+      mySnackbar(title: txtFailed, description: txtUnkownError);
+      return null;
+    } else {
+      mySnackbar(title: "Failed", description: "${res.body["message"]}");
     }
-  } else {
-    return emptyRes;
   }
-}
 
-Future<http.Response> addProjectApi(body) async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      var response = await client.post(
-        BASE_URL + PROJECT,
-        headers: {'Authorization': getToken(), "accept": "text/plain", "Content-Type": "application/json"},
-        body: jsonEncode(body),
-      );
-      print("Project Response: ${response.request.url}");
-      print("Project Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
+  static Future<Response?> updateProjectApi(body, id) async {
+    Response res = await httpService.putRequest(PROJECT + id, body).onError((error, stackTrace) => emptyRes);
+    if (res.statusCode == 200) {
+      return res;
+    } else if (res.statusCode == 700) {
+      mySnackbar(title: txtFailed, description: txtUnkownError);
+      return null;
+    } else {
+      mySnackbar(title: "Failed", description: "${res.body["message"]}");
     }
-  } else {
-    return emptyRes;
   }
-}
 
-Future<http.Response> updateProjectApi(body, id) async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      var response = await client.put(
-        BASE_URL + PROJECT + id.toString(),
-        headers: {'Authorization': getToken(), "accept": "text/plain", "Content-Type": "application/json"},
-        body: jsonEncode(body),
-      );
-      print("Project Response: ${response.request.url}");
-      print("Project Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
+  static Future<Response?> deleteProjectApi(id) async {
+    Response res = await httpService.deleteRequest(PROJECT + id).onError((error, stackTrace) => emptyRes);
+    if (res.statusCode == 200) {
+      return res;
+    } else if (res.statusCode == 700) {
+      mySnackbar(title: txtFailed, description: txtUnkownError);
+      return null;
+    } else {
+      mySnackbar(title: "Failed", description: "${res.body["message"]}");
     }
-  } else {
-    return emptyRes;
-  }
-}
-
-Future<http.Response> deleteProjectApi(id) async {
-  accessController.checkOnline();
-  if (accessController.isOnline.value) {
-    var client = http.Client();
-    try {
-      var response = await client.delete(
-        BASE_URL + PROJECT + id.toString(),
-        headers: {'Authorization': getToken()},
-      );
-      print("Project Response: ${response.request.url}");
-      print("Project Response: ${response.body}");
-      return response;
-    } catch (e) {
-      print(e);
-      return emptyRes;
-    } finally {
-      client.close();
-    }
-  } else {
-    return emptyRes;
   }
 }
