@@ -5,10 +5,10 @@ Color primaryColor = Colors.green;
 Color black = Colors.black;
 Color primaryGreyColor = Color(0xFFF1F9FD);
 Color grey = Colors.grey;
-Color greyDark = Colors.grey[700];
-Color greyLight = Colors.grey[300];
-Color grey200 = Colors.grey[200];
-Color grey100 = Colors.grey[100];
+Color? greyDark = Colors.grey[700];
+Color? greyLight = Colors.grey[300];
+Color? grey200 = Colors.grey.shade200;
+Color? grey100 = Colors.grey[100];
 Color textTitleColor = Color(0xff281856);
 Color lightTextColor = const Color(0x66281856);
 Color subTitleColor = const Color(0xff7a8aa5);
@@ -23,9 +23,9 @@ Color cardColor = Color(0xffE1EAF5);
 Color dashCardColor = Color(0xffF7F7FC);
 Color buttonClick = Color(0xff05404b);
 Color button = Colors.green;
-Color buttonColor = Colors.green[200];
+Color? buttonColor = Colors.green[200];
 Color locationColor = Color(0xff3366cc);
 Color inventoryColor = Color(0xff990099);
 Color boxesColor = Color(0xff109618);
 Color itemColor = Color(0xfffdbe19);
-List<BoxShadow> shadowList = [BoxShadow(color: Colors.grey[300], blurRadius: 30, offset: Offset(0, 10))];
+List<BoxShadow> shadowList = [BoxShadow(color: Colors.grey[300]!, blurRadius: 30, offset: Offset(0, 10))];
