@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'colors.dart';
 
@@ -11,12 +12,19 @@ import 'colors.dart';
 //   return res;
 // }
 
-mySnackbar({String title, String description}) {
-  return Get.snackbar(title, description, backgroundColor: grey, snackPosition: SnackPosition.BOTTOM);
+mySnackbar({required String title, required String description}) {
+  return Get.snackbar(title, description, backgroundColor: Colors.green.shade700, snackPosition: SnackPosition.BOTTOM, colorText: Colors.white);
 }
 
-Future<bool> snackBarBack({String title, String description}) async {
-  Get.snackbar(title, description, backgroundColor: grey, snackPosition: SnackPosition.BOTTOM, duration: Duration(seconds: 2));
+Future<bool> snackBarBack({required String title, required String description}) async {
+  Get.snackbar(
+    title,
+    description,
+    backgroundColor: Colors.green.shade700,
+    snackPosition: SnackPosition.BOTTOM,
+    colorText: Colors.white,
+    duration: Duration(seconds: 2),
+  );
   bool res = await Future.delayed(Duration(seconds: 2)).then((value) => true);
   return res;
 }
